284480 / Роли

+++ app/Http/Controllers/Admin/ArticlesController.php
@@ -9,6 +9,7 @@ use App\Contracts\Services\FlashMessageContract;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ArticleRequest;
 use App\Http\Requests\TagsRequest;
+use App\Models\Article;
 use App\Repositories\ArticlesRepository;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
@@ -23,6 +24,8 @@ class ArticlesController extends Controller
 
     public function index(): Factory|View|Application
     {
+        $this->authorize('viewAny', Article::class);
+
         $articles = $this->repository->findAll()->sortByDesc('updated_at');
         
         return view('pages.admin.articles.list', ['articles' => $articles]);
@@ -30,6 +33,8 @@ class ArticlesController extends Controller
 
     public function create(): Factory|View|Application
     {
+        $this->authorize('create', Article::class);
+
         return view('pages.admin.articles.create');
     }
 
@@ -39,6 +44,8 @@ class ArticlesController extends Controller
         FlashMessageContract $flashMessage,
         ArticleCreationServiceContract $createServise,
     ): RedirectResponse {
+        $this->authorize('create', Article::class);
+
         $fields = $request->validated();
         $tags = $tagsRequest->get('tags', []);
 
@@ -58,6 +65,8 @@ class ArticlesController extends Controller
     {
         $article = $this->repository->getById($id, ['image', 'tags']);
 
+        $this->authorize('update', $article);
+
         return view('pages.admin.articles.edit', ['article' => $article]);
     }
 
@@ -68,6 +77,8 @@ class ArticlesController extends Controller
         FlashMessageContract $flashMessage,
         ArticleUpdateServiceContract $updateServise,
     ): RedirectResponse {
+        $this->authorize('update', [Article::class, $id]);
+
         $fields = $request->validated();
         $tags = $tagsRequest->get('tags', []);
 
@@ -83,6 +94,8 @@ class ArticlesController extends Controller
         ArticleRemoveServiceContract $removeService,
         FlashMessageContract $flashMessage,
     ): RedirectResponse {
+        $this->authorize('delete', [Article::class, $id]);
+
         $removeService->delete($id);
 
         $flashMessage->success('Новость удалена');
